add check if port is in use

+++ package/kedro_viz/launchers/cli/run.py
@@ -116,6 +116,7 @@ def run(
         _PYPROJECT,
         _check_viz_up,
         _find_kedro_project,
+        _is_port_in_use,
         _start_browser,
         _wait_for,
         display_cli_message,
@@ -145,6 +146,18 @@ def run(
             "https://github.com/kedro-org/kedro-viz/releases.",
             "yellow",
         )
+
+    # Check if the port is already in use
+    if _is_port_in_use(host, port):
+        display_cli_message(
+            f"Error: Port {port} is already in use. Kedro Viz could not start.",
+            "red",
+        )
+        display_cli_message(
+            "Please specify a different port using the '--port' option.",
+            "red",
+        )
+        return
     try:
         if port in _VIZ_PROCESSES and _VIZ_PROCESSES[port].is_alive():
             _VIZ_PROCESSES[port].terminate()
